In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,7 +5,7 @@
   * Return: This Returns 0
   */
 
-int main(void)
+long int main(void)
 {
 	int i, n1 = 1, n2 = 2;
 	unsigned long next;
